igc: XDP packet TX timestamp
ADD the PTP timestamp of when a packet was transmitted to the XDP hints.
An application using AF_XDP can get this timestamp by inspecting the XDP
frame metadata when it gets to the completion queue.

Note that this implementation, with regard to the timestamp, is not
necessarily complete - its main purpose is to demostrate the XDP hints
API.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>