libbpf,xdp,igc: Use union to split XDP TX and RX metadata
Assuming that XDP metadata related to RX does not overlap with metadata
related to TX, this patch splits TX and RX data inside xdp_meta_generic
into different anonymous structs united by another anonymous union.

This allows more data to fit in the 32 bytes target size for
xdp_meta_generic. With help of BTF CO-RE, this is transparent for BPF
applications. For AF_XDP ones, libbpf has been modified to account for
the new layout.

The flip side is that holes appear in the struct, and one has to
carefully take care of padding, so that `btf_id` field is still at the
very end of struct.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>